requirements.txt
regex 
pandas 
numpy 
torch
faiss-cpu
openai 
pytest-asyncio
pytest-mock
python-multipart
tiktoken
typing-_extensions
datetime
uvicorn
fastapi 
pytest-env 
pycryptodome
pydantic
pydantic-settings 

bs4 
uvicorn 
urllib3 
transformers
python-dotenv
